ci: message handler redesigned update #70
