+++ src/message_handler/src/lib.rs
@@ -4,6 +4,8 @@ extern crate lazy_static;
 #[macro_use]
 extern crate log;
 
+pub mod constants;
 pub mod message_sender;
+pub mod message_traits;
 pub mod messages;
 pub mod node_messages;
