+++ src/message_handler/src/message_sender.rs
@@ -8,22 +8,22 @@ pub struct MessageSender {}
 
 impl MessageSender {
     pub fn send_block_msg(sender: &mut Sender<Option<MessageTypes>>, msg: SignedBlock) {
-        if let Ok(serialize_msg) = serialize(&NodeMessageTypes::SignedBlockEnum(msg)) {
-            let data = Some(MessageTypes::NodeMsg(serialize_msg));
-            let error: Result<(), TrySendError<Option<MessageTypes>>> = sender.try_send(data);
-            if error.is_err() {
-                error!("{:?}", error);
-            }
+        let data = Some(MessageTypes::NodeMsg(NodeMessageTypes::SignedBlockEnum(
+            msg,
+        )));
+        let error: Result<(), TrySendError<Option<MessageTypes>>> = sender.try_send(data);
+        if error.is_err() {
+            error!("{:?}", error);
         }
     }
 
     pub fn send_transaction_msg(sender: &mut Sender<Option<MessageTypes>>, msg: SignedTransaction) {
-        if let Ok(serialize_msg) = serialize(&NodeMessageTypes::SignedTransactionEnum(msg)) {
-            let data = Some(MessageTypes::NodeMsg(serialize_msg));
-            let error: Result<(), TrySendError<Option<MessageTypes>>> = sender.try_send(data);
-            if error.is_err() {
-                error!("{:?}", error);
-            }
+        let data = Some(MessageTypes::NodeMsg(
+            NodeMessageTypes::SignedTransactionEnum(msg),
+        ));
+        let error: Result<(), TrySendError<Option<MessageTypes>>> = sender.try_send(data);
+        if error.is_err() {
+            error!("{:?}", error);
         }
     }
 }
