+++ src/sdk/src/signed_transaction.rs
@@ -1,5 +1,3 @@
-use super::constants;
-use super::message_traits::Message;
 use std::convert::AsRef;
 use utils::serializer::{Deserialize, Serialize};
 
@@ -10,12 +8,4 @@ pub struct SignedTransaction {
     pub app_name: String,
     pub header: ::std::collections::HashMap<std::string::String, std::string::String>,
     pub signature: std::vec::Vec<u8>,
-}
-
-impl Message for SignedTransaction {
-    const TOPIC: &'static str = constants::NODE_MSG_TOPIC_STR[0];
-    const MODULE_TOPIC: &'static str = constants::NODE;
-    fn handler(&self) {
-        // info!("i am SignedTransaction handler");
-    }
 }
